split the coldfront.core.project.views module into multiple files for better organization. Refactor coldfront.core.project.views module by pulling out functions add comments to improve code readability split coldfront.core.project.utils into multiple utility files based on functionality fixed typo in project_detail.html template add SPDX headers to new utility and view files
